通信するクラスを追加
